baby-monitor.server/pi3/thermohygrometer/requirements.txt
smbus2==0.4.3
google-cloud-firestore==2.17.0
pytz==2024.1